# Refactor MinaCalc's Chisel function to improve performance