I profiled both the new and old calc to compare performance by deleting cache.db and timing from game-start to the first time I saw the menu screen using XCode's Instruments tool.

|Measurement|Old Calc | New Calc|
|---|---|---|
|Relative Chisel weight to total execution time|12%|10%|
|Total CPU-seconds spent in Chisel<br>(Slighty unreliable metric)|58.36|46.45|
|Total run time (CPU-Mins)<br>(Highly unreliable metric)|8.08|7.67|

Although this improvement is small, the new code is cleaner and easier to understand than the old code was, making this a change with no downsides.
There was zero difference in calculated values between the old calc and the new one with my changes, using the Calc test xml.
Related to #881 -- the first iteration is unimproved, but the 2nd-7th should take 50% less time on average.