@poco0317 2020-08-23
My scuffed profiling (run CalcEtternaMetadata 1000 times on the same file and measure the time taken)
Measured 3 times and took the average of the results.

|Song|438 Calc (ms) | 439 Calc (ms)|
|---|---|---|
|Game Time|19896.00|17424.33|
|Conspiracy of Silence|128951.00|116216.67|

Appears to be 10-15% faster.